add projectPageSwitch component
