+++ src/pages/project-page/project-page.tsx
@@ -1,30 +1,14 @@
-import { useParams } from "react-router-dom";
-import { lazy } from "react";
-
 import { ProjectNavMenu } from "../../components/project-nav-manu";
-import NotFoundPage from "../not-found-page";
+import { ProjectPageSwitch } from "./project-page-switch";
 
-const SummaryPage = lazy(() => import("../summary-page"));
-const DocumentListPage = lazy(() => import("../document-list-page"));
+
 
 export default function ProjectPage() {
-  const { location } = useParams();
-
-  const page = () => {
-    switch (location) {
-      case "document":
-        return <DocumentListPage />;
-      case "summary":
-        return <SummaryPage />;
-      default:
-        return <NotFoundPage />;
-    }
-  };
 
   return (
     <div>
       <ProjectNavMenu />
-      {page()}
+      <ProjectPageSwitch />
     </div>
   );
 }
